# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.StockBasicInfo:
    properties:
      area:
        example: 上海
        type: string
      created_at:
        example: "2025-08-13T10:00:00+08:00"
        type: string
      full_name:
        example: 上海浦东发展银行股份有限公司
        type: string
      id:
        example: 1
        type: integer
      industry:
        example: 银行
        type: string
      is_hs:
        example: true
        type: boolean
      listing_date:
        example: "1999-11-10"
        type: string
      market:
        example: 主板
        type: string
      name:
        example: 浦发银行
        type: string
      outstanding_shares:
        example: 2930026
        type: number
      status:
        example: 上市
        type: string
      symbol:
        example: "600000"
        type: string
      total_shares:
        example: 2930026
        type: number
      updated_at:
        example: "2025-08-13T10:00:00+08:00"
        type: string
    type: object
...
  models.StockIndicator:
    properties:
      boll_lower:
        type: number
      boll_mid:
        type: number
      boll_upper:
        type: number
      calc_date:
        description: 计算日期
        type: string
      created_at:
        type: string
      d:
        type: number
      dea:
        type: number
      dif:
        type: number
      id:
        type: integer
      j:
        type: number
      k:
        type: number
      ma5:
        type: number
      ma10:
        type: number
      ma20:
        type: number
      ma60:
        type: number
      macd:
        type: number
      rsi6:
        type: number
      rsi12:
        type: number
      rsi24:
        type: number
      symbol:
        description: 股票代码
        type: string
    type: object
...
host: 192.168.3.99:3003
info:
  contact: {}
  description: 股票基础信息管理 RESTful API
  title: 股票信息 API
  version: "1.0"
paths:
...
  /stock-indicators:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StockIndicator'
            type: array
      summary: 获取全部指标
      tags:
      - 股票指标
    post:
      consumes:
      - application/json
      parameters:
      - description: 股票指标数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StockIndicator'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StockIndicator'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 创建指标记录
      tags:
      - 股票指标
  /stock-indicators/{id}:
    delete:
      parameters:
      - description: 指标ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 删除指标
      tags:
      - 股票指标
    get:
      parameters:
      - description: 指标ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockIndicator'
        "404":
          description: Not Found
          schema:
            type: string
      summary: 根据ID获取指标
      tags:
      - 股票指标
    put:
      consumes:
      - application/json
      parameters:
      - description: 指标ID
        in: path
        name: id
        required: true
        type: integer
      - description: 股票指标数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StockIndicator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockIndicator'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 更新指标
      tags:
      - 股票指标
  /stock-indicators/symbol/{symbol}:
    get:
      parameters:
      - description: 股票代码
        in: path
        name: symbol
        required: true
        type: string
      - description: 计算日期(YYYY-MM-DD)
        in: query
        name: calc_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockIndicator'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 根据symbol和calc_date获取指标
      tags:
      - 股票指标
...
  /stocks:
    get:
      description: 获取所有股票的基础信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StockBasicInfo'
            type: array
      summary: 获取所有股票信息
      tags:
      - 股票信息
    post:
      consumes:
      - application/json
      description: 新增一条股票基础信息
      parameters:
      - description: 股票信息
        in: body
        name: stock
        required: true
        schema:
          $ref: '#/definitions/models.StockBasicInfo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StockBasicInfo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建股票信息
      tags:
      - 股票信息
  /stocks/{id}:
    delete:
      description: 根据ID删除股票记录
      parameters:
      - description: 股票ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除股票信息
      tags:
      - 股票信息
    get:
      description: 通过股票ID获取详细信息
      parameters:
      - description: 股票ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockBasicInfo'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取股票信息
      tags:
      - 股票信息
    put:
      consumes:
      - application/json
      description: 根据ID更新股票基础信息
      parameters:
      - description: 股票ID
        in: path
        name: id
        required: true
        type: integer
      - description: 股票信息
        in: body
        name: stock
        required: true
        schema:
          $ref: '#/definitions/models.StockBasicInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockBasicInfo'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新股票信息
      tags:
      - 股票信息
  /stocks/history:
    get:
      parameters:
      - description: 股票代码
        in: query
        name: symbol
        required: true
        type: string
      - description: 最近几天，默认7天
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: 获取股票最近 N 天历史数据（含技术指标和资金流）
      tags:
      - 股票综合信息
  /stocks/symbol/{symbol}:
    delete:
      description: 通过股票代码删除股票基础信息
      parameters:
      - description: 股票代码
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据股票代码删除股票信息
      tags:
      - 股票信息
    get:
      description: 通过股票代码获取股票基础信息
      parameters:
      - description: 股票代码
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StockBasicInfo'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据股票代码获取股票信息
      tags:
      - 股票信息
swagger: "2.0"